// arq-sdk/src/lib/interfaces/arq-list.interface.ts
export interface ArqList {
  /**
   * Value que tienen y que se utiliza para los guardados.
   */
  value: string;

  /**
   * Descripcion que se muestra en las listas, combos o autocomplete.
   */
  description: string;

  /**
   * Descripción que se muestra cuando el idioma es valenciano
   */
  descriptionV: string;
}