// arq-sdk/src/lib/components/arq-autocomplete/arq-autocomplete.service.d.ts
import { Observable } from 'rxjs';
import * as i0 from "@angular/core";
export declare class ArqAutocompleteService {
    constructor();
    _filter(options: any, value: string, lang: any): Observable<any>;
    static ɵfac: i0.ɵɵFactoryDeclaration<ArqAutocompleteService, never>;
    static ɵprov: i0.ɵɵInjectableDeclaration<ArqAutocompleteService>;
}